+++ extensions/telegram/src/bot-message-dispatch.draft-finalization.test.ts
@@ -21,232 +21,7 @@ import {
 } from "./bot-message-dispatch.test-harness.js";
 import type { TelegramMessageContext } from "./bot-message-dispatch.test-harness.js";
 
-describeTelegramDispatch("dispatchTelegramMessage fence-draft-finalization", () => {
-  it("keeps supersession latched when it arrives during adoption", async () => {
-    const sessionKey = "agent:main:telegram:direct:adoption-race";
-    let adoptionStarted: (() => void) | undefined;
-    const adoptionStartGate = new Promise<void>((resolve) => {
-      adoptionStarted = resolve;
-    });
-    let releaseAdoption: (() => void) | undefined;
-    const adoptionGate = new Promise<void>((resolve) => {
-      releaseAdoption = resolve;
-    });
-    dispatchReplyWithBufferedBlockDispatcher.mockImplementation(
-      async ({ dispatcherOptions, replyOptions }) => {
-        await replyOptions?.onTurnAdopted?.();
-        await dispatcherOptions.deliver({ text: "stale final" }, { kind: "final" });
-        return { queuedFinal: true };
-      },
-    );
-    deliverReplies.mockResolvedValue({ delivered: true });
-
-    const dispatchPromise = dispatchWithContext({
-      context: createContext({
-        ctxPayload: {
-          SessionKey: sessionKey,
-          ChatType: "direct",
-          MessageSid: "101",
-          RawBody: "long turn",
-          BodyForAgent: "long turn",
-          CommandBody: "long turn",
-          CommandAuthorized: true,
-        } as unknown as TelegramMessageContext["ctxPayload"],
-        msg: {
-          chat: { id: 123, type: "private" },
-          message_id: 101,
-          text: "long turn",
-        } as unknown as TelegramMessageContext["msg"],
-        chatId: 123,
-        isGroup: false,
-        threadSpec: { id: undefined, scope: "none" },
-      }),
-      streamMode: "off",
-      onTurnAdopted: async () => {
-        adoptionStarted?.();
-        await adoptionGate;
-      },
-    });
-    await adoptionStartGate;
-
-    const { supersedeTelegramReplyFence } = await import("./telegram-reply-fence.js");
-    expect(supersedeTelegramReplyFence(sessionKey)).toBe(true);
-    releaseAdoption?.();
-    await dispatchPromise;
-
-    expect(deliverReplies).not.toHaveBeenCalled();
-  });
-
-  it("lets authorized /stop kill an adopted run without the released fence controller", async () => {
-    const historyKey = "telegram:group:-100123";
-    const groupHistories = new Map([[historyKey, []]]);
-    // Core owns post-adoption abort via reply-run registry / handleStopCommand.
-    // Pin: after fence release, a core-owned abort still ends the run while the
-    // fence controller stays non-aborted.
-    const coreRunController = new AbortController();
-    let firstStarted: (() => void) | undefined;
-    const firstStartGate = new Promise<void>((resolve) => {
-      firstStarted = resolve;
-    });
-    let firstAbortSignal: AbortSignal | undefined;
-    let adoptTurn: (() => void | Promise<void>) | undefined;
-    let runSettled = false;
-    dispatchReplyWithBufferedBlockDispatcher
-      .mockImplementationOnce(async ({ replyOptions }) => {
-        firstAbortSignal = replyOptions?.abortSignal;
-        adoptTurn = replyOptions?.onTurnAdopted;
-        firstStarted?.();
-        await new Promise<void>((resolve) => {
-          const finish = () => {
-            if (runSettled) {
-              return;
-            }
-            runSettled = true;
-            resolve();
-          };
-          firstAbortSignal?.addEventListener("abort", finish, { once: true });
-          coreRunController.signal.addEventListener("abort", finish, { once: true });
-        });
-        return {
-          queuedFinal: false,
-          counts: { block: 0, final: 0, tool: 0 },
-        };
-      })
-      .mockImplementationOnce(async () => {
-        // Simulate core handleStopCommand / abortReplyRunBySessionId effect on the
-        // adopted registry-owned run (independent of the released fence controller).
-        coreRunController.abort();
-        return {
-          queuedFinal: false,
-          counts: { block: 0, final: 0, tool: 0 },
-        };
-      });
-
-    const createGroupContext = (messageId: number, body: string) =>
-      createContext({
-        ctxPayload: {
-          SessionKey: "agent:main:telegram:group:-100123",
-          ChatType: "group",
-          MessageSid: String(messageId),
-          RawBody: body,
-          BodyForAgent: body,
-          CommandBody: body,
-          CommandAuthorized: true,
-        } as unknown as TelegramMessageContext["ctxPayload"],
-        msg: {
-          chat: { id: -100123, type: "supergroup" },
-          message_id: messageId,
-          text: body,
-        } as unknown as TelegramMessageContext["msg"],
-        chatId: -100123,
-        isGroup: true,
-        historyKey,
-        historyLimit: 10,
-        groupHistories,
-        threadSpec: { id: undefined, scope: "none" },
-      });
-
-    const firstPromise = dispatchWithContext({
-      context: createGroupContext(99, "@bot long adopted turn"),
-      streamMode: "off",
-    });
-    await firstStartGate;
-    await adoptTurn?.();
-    expect(firstAbortSignal?.aborted).toBe(false);
-    expect(coreRunController.signal.aborted).toBe(false);
-
-    await dispatchWithContext({
-      context: createGroupContext(100, "/stop"),
-      streamMode: "off",
-    });
-
-    await firstPromise;
-    expect(firstAbortSignal?.aborted).toBe(false);
-    expect(coreRunController.signal.aborted).toBe(true);
-  });
-
-  it("keeps overlapping group deliveries non-superseded", async () => {
-    const historyKey = "telegram:group:-100123";
-    const groupHistories = new Map([[historyKey, []]]);
-    let firstStarted: (() => void) | undefined;
-    const firstStartGate = new Promise<void>((resolve) => {
-      firstStarted = resolve;
-    });
-    let releaseFirst: (() => void) | undefined;
-    const firstGate = new Promise<void>((resolve) => {
-      releaseFirst = resolve;
-    });
-    let secondStarted: (() => void) | undefined;
-    const secondStartGate = new Promise<void>((resolve) => {
-      secondStarted = resolve;
-    });
-    dispatchReplyWithBufferedBlockDispatcher
-      .mockImplementationOnce(async ({ dispatcherOptions, replyOptions }) => {
-        await replyOptions?.onTurnAdopted?.();
-        firstStarted?.();
-        await firstGate;
-        await dispatcherOptions.deliver({ text: "earlier group answer" }, { kind: "final" });
-        return {
-          queuedFinal: true,
-          counts: { block: 0, final: 1, tool: 0 },
-        };
-      })
-      .mockImplementationOnce(async ({ dispatcherOptions, replyOptions }) => {
-        await replyOptions?.onTurnAdopted?.();
-        secondStarted?.();
-        await dispatcherOptions.deliver({ text: "fresh group answer" }, { kind: "final" });
-        return {
-          queuedFinal: true,
-          counts: { block: 0, final: 1, tool: 0 },
-        };
-      });
-    deliverReplies.mockResolvedValue({ delivered: true });
-
-    const createGroupContext = (messageId: number, body: string) =>
-      createContext({
-        ctxPayload: {
-          SessionKey: "agent:main:telegram:group:-100123",
-          ChatType: "group",
-          MessageSid: String(messageId),
-          RawBody: body,
-          BodyForAgent: body,
-          CommandBody: body,
-          CommandAuthorized: true,
-        } as unknown as TelegramMessageContext["ctxPayload"],
-        msg: {
-          chat: { id: -100123, type: "supergroup" },
-          message_id: messageId,
-        } as unknown as TelegramMessageContext["msg"],
-        chatId: -100123,
-        isGroup: true,
-        historyKey,
-        historyLimit: 10,
-        groupHistories,
-        threadSpec: { id: undefined, scope: "none" },
-      });
-
-    const firstPromise = dispatchWithContext({
-      context: createGroupContext(99, "@bot first request"),
-      streamMode: "off",
-    });
-    await firstStartGate;
-    const secondPromise = dispatchWithContext({
-      context: createGroupContext(100, "@bot second request"),
-      streamMode: "off",
-    });
-    await secondStartGate;
-    releaseFirst?.();
-    await Promise.all([firstPromise, secondPromise]);
-
-    const deliveredTexts = deliverReplies.mock.calls.flatMap((call) =>
-      ((call[0] as { replies?: Array<{ text?: string }> }).replies ?? []).map(
-        (reply) => reply.text,
-      ),
-    );
-    expect(deliveredTexts).toContain("earlier group answer");
-    expect(deliveredTexts).toContain("fresh group answer");
-  });
-
+describeTelegramDispatch("dispatchTelegramMessage draft-finalization", () => {
   it("does not drop any long-final text after a generic lane rotation", async () => {
     const { answerDraftStream } = setupDraftStreams({ answerMessageId: 2001 });
     dispatchReplyWithBufferedBlockDispatcher.mockImplementation(
